refactor family utils
