+++ src/components/Sidebar/SubscriptionCard/contentOverlay/index.tsx
@@ -1,7 +1,7 @@
 import { IndividualSubscriptionCardContent } from './individual';
 import { FamilySubscriptionCardContent } from './family';
 import React from 'react';
-import { hasNonAdminFamilyMembers } from 'utils/billing';
+import { hasNonAdminFamilyMembers } from 'utils/user/family';
 import { Overlay, SpaceBetweenFlex } from 'components/Container';
 import { UserDetails } from 'types/user';
